update gitignore, move existing code to this directory
