+++ SG3_Devices/connect/.gitignore
@@ -4,3 +4,4 @@ config/.env
 logs/
 __pycache__/
 *.pyc
+.idea
