--- SG3_Devices/connect/config/.env.example
@@ -1,1 +1,0 @@
-Fill in env
