+++ SG3_Devices/connect/gateway/config/.env.example
