--- SG3_Devices/connect/gateway/src/gateway.py
@@ -1,1 +1,0 @@
-# Gateway Code
