+++ SG3_Devices/connect/gateway/src/local_api.py
