+++ SG3_Devices/connect/gateway/src/serial_client.py
